pyproject.toml
[project]
name = "bilicast"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Fatpandac", email = "i@fatpandac.com" }
]
dependencies = [
    "bilix>=0.18.8",
    "feedgen>=1.0.0",
    "fastapi>=0.111.0",
    "pyyaml>=6.0.1",
    "uvicorn>=0.30.1",
    "apscheduler>=3.10.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye.scripts]
dev = "uvicorn --host 0.0.0.0 --reload src.main:app"
test = "pytest --cov=src --cov-report=html --cov-fail-under=90"

[project.scripts]
hello = "bilicast:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.2.2",
    "pytest-cov>=5.0.0",
    "pytest-sugar>=1.0.0",
    "pre-commit>=3.5.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/bilicast"]
